Fix Layout - Fragment Cart - Fragment Product Detail - Fragment User Info - Fix error can not return home from product detail - Rename project name
