+++ app/google-services.json
@@ -10,7 +10,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:424518794109:android:cfc9aa5c88daf444bfaa14",
         "android_client_info": {
-          "package_name": "com.example.se171889project"
+          "package_name": "com.example.bookstore"
         }
       },
       "oauth_client": [],
