+++ app/src/main/java/com/example/bookstore/HomeActivity.java
@@ -1,0 +1,59 @@
+package com.example.bookstore;
+
+import android.os.Bundle;
+
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.navigation.NavController;
+import androidx.navigation.Navigation;
+import androidx.navigation.ui.AppBarConfiguration;
+import androidx.navigation.ui.NavigationUI;
+
+import com.example.bookstore.databinding.ActivityHomeBinding;
+
+public class HomeActivity extends AppCompatActivity {
+
+    private ActivityHomeBinding binding;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        binding = ActivityHomeBinding.inflate(getLayoutInflater());
+        setContentView(binding.getRoot());
+
+        BottomNavigationView navView = findViewById(R.id.nav_view);
+        // Passing each menu ID as a set of Ids because each
+        // menu should be considered as top level destinations.
+        AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(
+                R.id.navigation_home, R.id.navigation_cart, R.id.navigation_payment_history, R.id.navigation_user_info)
+                .build();
+        NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_activity_home);
+        NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
+        NavigationUI.setupWithNavController(binding.navView, navController);
+        navView.setOnNavigationItemSelectedListener(item -> {
+            if (item.getItemId() == R.id.navigation_home) {
+                navController.popBackStack(R.id.navigation_home, false); // Trở về HomeFragment mà không thêm vào ngăn xếp
+                return true;
+            } else if (item.getItemId() == R.id.navigation_cart) {
+                navController.navigate(R.id.navigation_cart);
+                return true;
+            } else if (item.getItemId() == R.id.navigation_payment_history) {
+                navController.navigate(R.id.navigation_payment_history);
+                return true;
+            } else if (item.getItemId() == R.id.navigation_user_info) {
+                navController.navigate(R.id.navigation_user_info);
+                return true;
+            }
+            return false;
+        });
+    }
+
+    @Override
+    public boolean onSupportNavigateUp() {
+        NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_activity_home);
+        return NavigationUI.navigateUp(navController, new AppBarConfiguration.Builder(navController.getGraph()).build())
+                || super.onSupportNavigateUp();
+    }
+}
