+++ app/src/main/java/com/example/bookstore/SignupActivity.java
@@ -1,0 +1,80 @@
+package com.example.bookstore;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.text.TextUtils;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.TextView;
+import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+
+public class SignupActivity extends AppCompatActivity {
+    EditText editTextUsername, editTextPassword;
+    Button signUp;
+    TextView signIn;
+    FirebaseAuth firebaseAuth = FirebaseAuth.getInstance();
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_signup);
+
+        editTextUsername = findViewById(R.id.username_input);
+        editTextPassword = findViewById(R.id.password_input);
+        signIn = findViewById(R.id.signin_textview);
+        signUp = findViewById(R.id.signup_btn);
+
+        signIn.setOnClickListener(v -> {
+            Intent intent = new Intent(SignupActivity.this, MainActivity.class);
+            startActivity(intent);
+        });
+
+        signUp.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                String username = String.valueOf(editTextUsername.getText());
+                String password = String.valueOf(editTextPassword.getText());
+
+                if (TextUtils.isEmpty(username)) {
+                    Toast.makeText(SignupActivity.this, "Enter Username", Toast.LENGTH_SHORT).show();
+                    return;
+                }
+                if (!android.util.Patterns.EMAIL_ADDRESS.matcher(username).matches()) {
+                    Toast.makeText(SignupActivity.this, "Enter a valid email address", Toast.LENGTH_SHORT).show();
+                    return;
+                }
+                if (TextUtils.isEmpty(password)) {
+                    Toast.makeText(SignupActivity.this, "Enter Password", Toast.LENGTH_SHORT).show();
+                    return;
+                }
+
+                String hashedPassword = PasswordUtils.hashPassword(password);
+
+                firebaseAuth.createUserWithEmailAndPassword(username, hashedPassword)
+                        .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
+                            @Override
+                            public void onComplete(@NonNull Task<AuthResult> task) {
+                                if(task.isSuccessful()) {
+                                    Toast.makeText(SignupActivity.this, "Sign Up Successfully", Toast.LENGTH_SHORT).show();
+                                    Intent intent = new Intent(SignupActivity.this, MainActivity.class);
+                                    startActivity(intent);
+                                    finish();
+                                }
+                                else {
+                                    Toast.makeText(SignupActivity.this, "Sign Up Failed", Toast.LENGTH_SHORT).show();
+                                }
+                            }
+                        });
+            }
+        });
+    }
+}
