+++ app/src/main/java/com/example/bookstore/model/Product.java
@@ -1,0 +1,65 @@
+package com.example.bookstore.model;
+
+import java.io.Serializable;
+
+public class Product implements Serializable {
+    private String id;
+    private String name;
+    private double price;
+    private String imageUrl;
+    private String description;
+
+    public Product() {
+        // Default constructor required for Firebase
+    }
+
+    public Product(String id, String name, double price, String imageUrl, String description) {
+        this.id = id;
+        this.name = name;
+        this.price = price;
+        this.imageUrl = imageUrl;
+        this.description = description;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public double getPrice() {
+        return price;
+    }
+
+    public void setPrice(double price) {
+        this.price = price;
+    }
+
+    public String getImageUrl() {
+        return imageUrl;
+    }
+
+    public void setImageUrl(String imageUrl) {
+        this.imageUrl = imageUrl;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    // Getters and Setters
+}
