+++ app/src/main/java/com/example/bookstore/model/UserEntity.java
@@ -1,0 +1,39 @@
+package com.example.bookstore.model;
+
+public class UserEntity {
+    private String uid;
+    private String email;
+    private String name;
+
+    public UserEntity() {
+    }
+    public UserEntity(String uid, String email, String name) {
+        this.uid = uid;
+        this.email = email;
+        this.name = name;
+    }
+
+    public String getUid() {
+        return uid;
+    }
+
+    public void setUid(String uid) {
+        this.uid = uid;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
