+++ app/src/main/java/com/example/bookstore/ui/cart/MapActivity.java
@@ -1,0 +1,63 @@
+package com.example.bookstore.ui.cart;
+
+import android.content.Intent;
+import android.os.Bundle;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.bookstore.R;
+import com.google.android.gms.maps.CameraUpdateFactory;
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.OnMapReadyCallback;
+import com.google.android.gms.maps.SupportMapFragment;
+import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.MarkerOptions;
+
+public class MapActivity extends AppCompatActivity implements OnMapReadyCallback {
+
+    private GoogleMap mMap;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_map);
+
+        SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager()
+                .findFragmentById(R.id.map);
+        assert mapFragment != null;
+        mapFragment.getMapAsync(this);
+    }
+
+    @Override
+    public void onMapReady(@NonNull GoogleMap googleMap) {
+        mMap = googleMap;
+
+//        if (ActivityCompat.checkSelfPermission(this, android.Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(this, android.Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+//            // TODO: Consider calling
+//            //    ActivityCompat#requestPermissions
+//            // here to request the missing permissions, and then overriding
+//            //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
+//            //                                          int[] grantResults)
+//            // to handle the case where the user grants the permission. See the documentation
+//            // for ActivityCompat#requestPermissions for more details.
+//            return;
+//        }
+//        mMap.setMyLocationEnabled(true); // Bật hiển thị vị trí của người dùng
+        // Thiết lập vị trí mặc định
+        LatLng defaultLocation = new LatLng(10.826676948606515, 106.60568896196074);
+        mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(defaultLocation, 10));
+
+        // Thêm sự kiện nhấp vào bản đồ
+        mMap.setOnMapClickListener(latLng -> {
+            mMap.clear(); // Xóa các marker cũ
+            mMap.addMarker(new MarkerOptions().position(latLng)); // Thêm marker tại vị trí nhấp
+            // Trả về vị trí đã chọn
+            Intent returnIntent = new Intent();
+            returnIntent.putExtra("latitude", latLng.latitude);
+            returnIntent.putExtra("longitude", latLng.longitude);
+            setResult(RESULT_OK, returnIntent);
+            finish(); // Đóng activity
+        });
+    }
+}
