--- app/src/main/java/com/example/se171889project/PasswordUtils.java
@@ -1,23 +1,0 @@
-package com.example.se171889project;
-
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-
-public class PasswordUtils {
-    public static String hashPassword(String password) {
-        try {
-            MessageDigest md = MessageDigest.getInstance("SHA-256");
-            md.update(password.getBytes());
-            byte[] bytes = md.digest();
-            StringBuilder sb = new StringBuilder();
-            for (byte b : bytes) {
-                sb.append(String.format("%02x", b));
-            }
-            return sb.toString();
-        } catch (NoSuchAlgorithmException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-}
