--- app/src/main/java/com/example/se171889project/model/CartItem.java
@@ -1,64 +1,0 @@
-package com.example.se171889project.model;
-// CartItem.java
-public class CartItem {
-    private String productId;
-    private String name;
-    private double price;
-    private int quantity;
-    private String imageUrl;
-
-    public CartItem() {
-        // Constructor rỗng cần thiết cho Firebase
-    }
-
-    public CartItem(String productId, String name, double price, int quantity, String imageUrl) {
-        this.productId = productId;
-        this.name = name;
-        this.price = price;
-        this.quantity = quantity;
-        this.imageUrl = imageUrl;
-    }
-
-    public String getProductId() {
-        return productId;
-    }
-
-    public void setProductId(String productId) {
-        this.productId = productId;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-    public void setPrice(double price) {
-        this.price = price;
-    }
-
-    public int getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(int quantity) {
-        this.quantity = quantity;
-    }
-
-    public String getImageUrl() {
-        return imageUrl;
-    }
-
-    public void setImageUrl(String imageUrl) {
-        this.imageUrl = imageUrl;
-    }
-
-    // Getters and setters
-    // ...
-}
