+++ app/src/main/res/drawable/rectangle_button.xml
@@ -1,0 +1,3 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <corners android:radius="0dp" /> <!-- Sharp corners -->
+</shape>
