+++ app/src/main/res/layout/fragment_product_detail.xml
@@ -5,52 +5,104 @@
     android:layout_height="match_parent"
     android:padding="16dp">
 
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal">
+
+        <TextView
+            android:layout_width="160dp"
+            android:layout_height="wrap_content"
+            android:textSize="24sp"
+            android:textStyle="bold"
+            android:text="Tên: "
+            android:layout_weight="1"/>
+
+        <TextView
+            android:id="@+id/product_name_detail"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textSize="24sp"
+            android:layout_weight="1"/>
+
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        android:weightSum="10">
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textSize="24sp"
+            android:textStyle="bold"
+            android:text="Giá: "
+            android:layout_weight="1"/>
+
+        <TextView
+            android:id="@+id/product_price_detail"
+            android:layout_width="wrap_content"
+            android:layout_height="match_parent"
+            android:layout_weight="9"
+            android:textColor="#f0a235"
+            android:textSize="20sp" />
+
+    </LinearLayout>
+
     <ImageView
         android:id="@+id/product_image_detail"
         android:layout_width="match_parent"
         android:layout_height="200dp"
         android:scaleType="centerCrop"
-        android:src="@drawable/ic_launcher_background" />
-
-    <TextView
-        android:id="@+id/product_name_detail"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textSize="24sp"
-        android:textColor="@android:color/black"
+        android:src="@drawable/ic_launcher_background"
         android:layout_marginTop="8dp"/>
 
-    <TextView
-        android:id="@+id/product_price_detail"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textSize="20sp"
-        android:textColor="@android:color/holo_green_dark"
-        android:layout_marginTop="4dp"/>
-
-    <TextView
-        android:id="@+id/product_description"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textSize="16sp"
-        android:textColor="@android:color/darker_gray"
-        android:layout_marginTop="8dp"/>
-
-    <!-- EditText để chọn số lượng -->
-    <EditText
-        android:id="@+id/quantity_edit_text"
+    <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:hint="Nhập số lượng"
-        android:inputType="number"
-        android:layout_marginTop="16dp"/>
+        android:orientation="vertical">
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="200dp"
+            android:orientation="horizontal"
+            android:layout_weight="1">
+
+            <TextView
+                android:layout_width="150dp"
+                android:layout_height="wrap_content"
+                android:textSize="24sp"
+                android:textStyle="bold"
+                android:text="Mô tả: "
+                android:layout_weight="1"/>
+
+            <TextView
+                android:id="@+id/product_description"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textSize="16sp"
+                android:textColor="@android:color/darker_gray"
+                android:layout_weight="1"/>
+        </LinearLayout>
+
+        <!-- EditText để chọn số lượng -->
+        <EditText
+            android:id="@+id/quantity_edit_text"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:hint="Nhập số lượng"
+            android:layout_weight="1"
+            android:inputType="number"/>
+    </LinearLayout>
 
     <!-- Nút thêm vào giỏ hàng -->
     <Button
         android:id="@+id/add_to_cart_button"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Thêm vào giỏ hàng"
-        android:layout_marginTop="16dp"/>
+        android:layout_marginTop="16dp"
+        android:background="@drawable/rectangle_button"
+        android:text="Thêm vào giỏ hàng"/>
 
 </LinearLayout>
