+++ app/src/main/res/layout/fragment_user_info.xml
@@ -21,40 +21,45 @@
         android:textSize="16sp"
         android:layout_marginBottom="16dp"/>
 
-    <Button
-        android:id="@+id/btn_logout"
+    <com.google.android.material.circularreveal.CircularRevealLinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Đăng xuất"
-        android:layout_marginBottom="8dp"/>
+        android:orientation="horizontal">
 
-    <Button
-        android:id="@+id/btn_change_password"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Đổi mật khẩu"
-        android:layout_marginBottom="16dp"/>
+        <Button
+            android:id="@+id/btn_logout"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:text="Đăng xuất"
+            android:backgroundTint="@color/btn_logut_bg"
+            android:layout_marginBottom="8dp"/>
+
+        <Button
+            android:id="@+id/btn_change_password"
+            android:layout_weight="1"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Đổi mật khẩu"
+            android:backgroundTint="@color/design_default_color_primary_dark"
+            android:layout_marginBottom="16dp"/>
+
+    </com.google.android.material.circularreveal.CircularRevealLinearLayout>
 
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Về cửa hàng"
+        android:text="Thông tin cửa hàng"
         android:textSize="18sp"
         android:textStyle="bold"
         android:layout_marginBottom="8dp"/>
-
-    <fragment
-        android:id="@+id/map_fragment"
-        android:name="com.google.android.gms.maps.SupportMapFragment"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"/>
 
     <!-- Thêm TextView để hiển thị tên và địa chỉ cửa hàng -->
     <TextView
         android:id="@+id/tv_store_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Tên cửa hàng: Watch Store FPT"
+        android:text="Tên: Book Store FPT"
         android:textSize="16sp"
         android:layout_marginTop="8dp"
         android:layout_marginBottom="4dp"/>
@@ -63,8 +68,14 @@
         android:id="@+id/tv_store_location"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Địa chỉ: Nhà Văn hóa Sinh viên TP.HCM, tầng 4 phòng 409"
+        android:text="Địa chỉ: Nhà Văn hóa Sinh viên TP.HCM, tầng 4 phòng 407"
         android:textSize="16sp"
         android:layout_marginBottom="16dp"/>
+
+    <fragment
+        android:id="@+id/map_fragment"
+        android:name="com.google.android.gms.maps.SupportMapFragment"
+        android:layout_width="match_parent"
+        android:layout_height="200dp"/>
 
 </LinearLayout>
