+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="com.example.se171889project.ui.home.HomeFragment"
+        android:name="com.example.bookstore.ui.home.HomeFragment"
         android:label="@string/title_home"
         tools:layout="@layout/fragment_home" >
         <action
@@ -26,7 +26,7 @@
 
     <fragment
         android:id="@+id/navigation_cart"
-        android:name="com.example.se171889project.ui.cart.CartFragment"
+        android:name="com.example.bookstore.ui.cart.CartFragment"
         android:label="@string/title_cart"
         tools:layout="@layout/fragment_cart" >
         <action
@@ -42,7 +42,7 @@
 
     <fragment
         android:id="@+id/navigation_payment_history"
-        android:name="com.example.se171889project.ui.payment.PaymentHistoryFragment"
+        android:name="com.example.bookstore.ui.payment.PaymentHistoryFragment"
         android:label="@string/title_payment_history"
         tools:layout="@layout/fragment_payment_history" >
         <action
@@ -58,7 +58,7 @@
 
     <fragment
         android:id="@+id/navigation_user_info"
-        android:name="com.example.se171889project.ui.user.UserInfoFragment"
+        android:name="com.example.bookstore.ui.user.UserInfoFragment"
         android:label="@string/title_user_info"
         tools:layout="@layout/fragment_user_info" >
         <action
@@ -73,7 +73,7 @@
     </fragment>
     <fragment
         android:id="@+id/productDetailFragment"
-        android:name="com.example.se171889project.ui.productdetail.ProductDetailFragment"
+        android:name="com.example.bookstore.ui.productdetail.ProductDetailFragment"
         android:label="Product Detail" >
         <action
             android:id="@+id/action_productDetailFragment_to_navigation_home"
